+++ tests/test_models/test_mutables/group/test_mutable_channel_groups.py
@@ -8,7 +8,7 @@ import torch.nn as nn
 from mmrazor.models.architectures.dynamic_ops.mixins import DynamicChannelMixin
 from mmrazor.models.mutables.mutable_channel import (
     L1MutableChannelGroup, MutableChannelGroup, SequentialMutableChannelGroup)
-from mmrazor.models.mutables.mutable_channel.groups.channel_group import (  # noqa
+from mmrazor.models.mutables.mutable_channel.units.channel_group import (  # noqa
     Channel, ChannelGroup)
 from mmrazor.structures.graph import ModuleGraph as ModuleGraph
 from ....data.models import LineModel
